Fixed duplicate states and moved Dialog to separate file.
